clean up DataLogger
