+++ aflac2020/Makefile.inc
@@ -8,7 +8,6 @@ Observer.o \
 Navigator.o \
 LineTracer.o \
 BlindRunner.o \
-DataLogger.o \
 utility.o
 
 SRCLANG := c++
