+++ aflac2020/app.cfg
@@ -22,5 +22,4 @@ ATT_MOD("Observer.o");
 ATT_MOD("Navigator.o");
 ATT_MOD("LineTracer.o");
 ATT_MOD("BlindRunner.o");
-ATT_MOD("DataLogger.o");
 ATT_MOD("utility.o");
